+++ src/openapi/finatic_server/models/success_payload_dict_str_any.py
@@ -21,21 +21,18 @@ from pydantic import BaseModel, ConfigDict, Field, StrictStr
 from typing import Any, ClassVar, Dict, List, Optional
 from typing import Optional, Set
 from typing_extensions import Self
-from pydantic_core import to_jsonable_python
 
 class SuccessPayloadDictStrAny(BaseModel):
     """
     SuccessPayloadDictStrAny
     """ # noqa: E501
     id: Optional[StrictStr] = Field(default=None, alias="_id")
-    data: Optional[Dict[str, Any]] = Field(default=None, description="The response data (None when error is present)")
-    meta: Optional[Dict[str, Any]] = Field(default=None, description="Optional metadata (pagination, etc.)")
-    additional_properties: Dict[str, Any] = {}
+    data: Optional[Dict[str, Any]] = None
+    meta: Optional[Dict[str, Any]] = None
     __properties: ClassVar[List[str]] = ["_id", "data", "meta"]
 
     model_config = ConfigDict(
-        validate_by_name=True,
-        validate_by_alias=True,
+        populate_by_name=True,
         validate_assignment=True,
         protected_namespaces=(),
     )
@@ -47,7 +44,8 @@ class SuccessPayloadDictStrAny(BaseModel):
 
     def to_json(self) -> str:
         """Returns the JSON representation of the model using alias"""
-        return json.dumps(to_jsonable_python(self.to_dict()))
+        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
+        return json.dumps(self.to_dict())
 
     @classmethod
     def from_json(cls, json_str: str) -> Optional[Self]:
@@ -63,10 +61,8 @@ class SuccessPayloadDictStrAny(BaseModel):
         * `None` is only added to the output dict for nullable fields that
           were set at model initialization. Other fields with value `None`
           are ignored.
-        * Fields in `self.additional_properties` are added to the output dict.
         """
         excluded_fields: Set[str] = set([
-            "additional_properties",
         ])
 
         _dict = self.model_dump(
@@ -74,11 +70,6 @@ class SuccessPayloadDictStrAny(BaseModel):
             exclude=excluded_fields,
             exclude_none=True,
         )
-        # puts key-value pairs in additional_properties in the top level
-        if self.additional_properties is not None:
-            for _key, _value in self.additional_properties.items():
-                _dict[_key] = _value
-
         # set to None if data (nullable) is None
         # and model_fields_set contains the field
         if self.data is None and "data" in self.model_fields_set:
@@ -105,11 +96,4 @@ class SuccessPayloadDictStrAny(BaseModel):
             "data": obj.get("data"),
             "meta": obj.get("meta")
         })
-        # store additional fields in additional_properties
-        for _key in obj.keys():
-            if _key not in cls.__properties:
-                _obj.additional_properties[_key] = obj.get(_key)
-
         return _obj
-
-